fix: Safely execute DOM bindings in legacy modules
- Wrapped `initTrackerBindings()` and `bindCoachView()` in `document.readyState` checks to prevent execution errors when modules load before the DOM is ready.
- Extracted heavy event handler logic from `bindCoachView` into `legacy/modules/coachUtils.js` to ensure the `bindCoachView` function complies with the 80-line limit rule.
- Maintained existing functionality and dependencies via dependency injection for the extracted `handleCoachViewClick` utility.
